github-accounts-analyzer/requirements.txt
# confluent-kafka[avro,json,protobuf]>=1.4.2
confluent-kafka
loguru
PyGithub